refactor: 목표 무한 스크롤 로직 hooks로 분리
